Refactor namespace for BackupService
