+++ Jellyfin.Api/Controllers/BackupController.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using Jellyfin.Server.Implementations.Backup;
+using Jellyfin.Server.Implementations.SystemBackupService;
 using MediaBrowser.Common.Api;
 using MediaBrowser.Controller.SystemBackupService;
 using Microsoft.AspNetCore.Authorization;
