+++ tests/Jellyfin.Api.Tests/Controllers/SystemControllerTests.cs
@@ -1,5 +1,5 @@
 using Jellyfin.Api.Controllers;
-using Jellyfin.Server.Implementations.Backup;
+using Jellyfin.Server.Implementations.SystemBackupService;
 using MediaBrowser.Common.Net;
 using MediaBrowser.Controller;
 using MediaBrowser.Model.IO;
